fix npm test due to the moving parts
