+++ .storybook/config.js
@@ -61,7 +61,6 @@ function loadStories() {
   require('../stories/AutoCompleteComposite');
   require('../stories/GoogleAddressInputWithLabel');
   require('../stories/BackOffice/Button/CompositeStory');
-  require('../stories/Viewer/Button/CompositeStory');
   require('../stories/Breadcrubms/CompositeStory');
   require('../stories/Tabs');
   require('../stories/Tooltip/Composite/CompositeStory');
