--- stories/Viewer/Button/ButtonStandard.js
@@ -1,112 +1,0 @@
-import React, {Component, PropTypes} from 'react';
-
-import Template from './Template';
-import RadioGroup from '../../../src/RadioGroup';
-import Label from '../../../src/Label';
-import Input from '../../../src/Input';
-import ToggleSwitch from '../../../src/ToggleSwitch';
-import IconChooser from './IconChooser';
-import * as Icons from '../../../src/Icons/dist';
-
-import styles from './ExampleButton.scss';
-
-class ButtonStandard extends Component {
-
-  static propTypes = {
-    onChange: PropTypes.func
-  };
-
-  state = {
-    disabled: false,
-    theme: 'fullblue',
-    text: 'Click On Me',
-    height: 'medium'
-  };
-
-  render() {
-    return (
-      <from className={styles.form}>
-        <div className={styles.input}>
-
-          <div className={styles.option}>
-            <Label>Type</Label>
-            <div className={styles.flex}>
-              <RadioGroup
-                display="horizontal"
-                value={this.state.theme}
-                onChange={theme => this.setState({theme})}
-              >
-                <RadioGroup.Radio value="fullblue">Primary</RadioGroup.Radio>
-                <RadioGroup.Radio value="transparentblue">Secondary</RadioGroup.Radio>
-                <RadioGroup.Radio value="whiteblue">Secondary on grey</RadioGroup.Radio>
-              </RadioGroup>
-            </div>
-          </div>
-
-          <div className={styles.option}>
-            <Label>Disabled</Label>
-            <div className={styles.flex}>
-              <ToggleSwitch
-                size="small"
-                checked={this.state.disabled}
-                onChange={() => this.setState({disabled: !this.state.disabled})}
-              />
-            </div>
-          </div>
-
-          <div className={styles.option}>
-            <Label>Prefix Icon</Label>
-            <div className={styles.flex}>
-              <IconChooser
-                onSelect={option => this.setState({prefixIcon: React.createElement(Icons[option.id])})}
-              />
-            </div>
-          </div>
-
-          <div className={styles.option}>
-            <Label>Suffix Icon</Label>
-            <div className={styles.flex}>
-              <IconChooser
-                onSelect={option => this.setState({suffixIcon: React.createElement(Icons[option.id])})}
-              />
-            </div>
-          </div>
-
-          <div className={styles.option}>
-            <Label>Text</Label>
-            <div className={styles.flex}>
-              <Input
-                size="small"
-                value={this.state.text}
-                onChange={e => this.setState({text: e.target.value})}
-              />
-            </div>
-          </div>
-
-          <div className={styles.option}>
-            <Label>Size</Label>
-            <div className={styles.flex}>
-              <RadioGroup
-                display="horizontal"
-                value={this.state.height}
-                onChange={height => this.setState({height})}
-              >
-                <RadioGroup.Radio value="small">Small</RadioGroup.Radio>
-                <RadioGroup.Radio value="medium">Regular</RadioGroup.Radio>
-                <RadioGroup.Radio value="large">Large</RadioGroup.Radio>
-              </RadioGroup>
-            </div>
-          </div>
-        </div>
-
-        <div className={styles[this.state.theme === 'whiteblue' ? 'output-lightblue' : 'output']}>
-          <div className={`${styles[this.state.theme]} ${styles.exampleWrapper}`}>
-            <Template {...this.state} onChange={this.props.onChange}/>
-          </div>
-        </div>
-      </from>
-    );
-  }
-}
-
-export default ButtonStandard;
