--- stories/Viewer/Button/ExampleWithIcons.js
@@ -1,34 +1,0 @@
-import React, {Component} from 'react';
-import {Button} from 'wix-style-react/Viewer';
-import {Close} from '../../../src/Icons/dist';
-
-const style = {
-  display: 'inline-block',
-  padding: '0 5px',
-  width: '140px',
-  lineHeight: '22px'
-};
-
-class IconsExample extends Component {
-  render() {
-    return (
-      <div>
-        <div className="ltr" style={style}>
-          Prefix<br/>
-          <Button id="button" dataHook="story-button-prefix" prefixIcon={<Close/>}>Prefix</Button>
-        </div>
-
-        <div className="ltr" style={style}>
-          Suffix<br/>
-          <Button id="button" dataHook="story-button-suffix" suffixIcon={<Close/>}>Suffix</Button>
-        </div>
-      </div>
-    );
-  }
-}
-
-export default () =>
-  <IconsExample/>;
-
-
-
