+++ testkit/index.js
@@ -45,6 +45,9 @@ export const checkboxTestkitFactory = testkitFactoryCreator(checkboxDriverFactor
 import buttonDriverFactory from '../src/Backoffice/Button/Button.driver';
 export const buttonTestkitFactory = testkitFactoryCreator(buttonDriverFactory);
 
+import viewerButtonDriverFactory from '../src/Backoffice/Button/Button.driver';
+export const viewerButtonTestkitFactory = testkitFactoryCreator(viewerButtonDriverFactory);
+
 import textLinkDriverFactory from '../src/BaseComponents/TextLink/TextLink.driver';
 export const textLinkTestkitFactory = testkitFactoryCreator(textLinkDriverFactory);
